#40 주문에 옵션추가 (#43)
* - 메뉴 삭제 기능 추가
- 메뉴 유효성 추가
    - 메뉴 id가 존하는지 체크
    - 메뉴 id와 가게 id가 존재하는지 체크

* #20 잘못된 테스트 수정

* pick 5dc1c46 #20 메뉴 삭제
    pick 064a2b0 #21 메뉴 업데이트

* * #27 옵션그룹등록

* #27 옵션그룹 목록 조회

- 조건은 가게 ID

* #27 옵션그룹 삭제

- 옵션그룹이 존하는지 검증하는 기능 추가
- 다른 가게의 옵션그룹인지 검증하는 기능 추가

* #27 옵션그룹 업데이트

* #27 옵션그룹등록

* #31 옵션등록

* #31 옵션수정

* #31 옵션 삭제

* #31 옵션 상세조회

* #27 옵션그룹등록

* #27 옵션그룹등록

* #32 옵션그룹과 옵션들 조회

* #35 옵션그룹들 삭제후 등록

- 메뉴옵션그룹들 삭제 후 메뉴옵션그룹들 저장
- 메뉴옵션그룹들 중복 미허용

* #35 메뉴옵션그룹들 조회

* #39 세션 => ownerDetails

* #40 주문에 옵션추가

+++ src/main/java/com/flab/commerce/domain/menu/Menu.java
@@ -1,7 +1,8 @@
 package com.flab.commerce.domain.menu;
 
+import com.flab.commerce.domain.menuoptiongroup.MenuOptionGroup;
 import com.flab.commerce.exception.BadInputException;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.ZonedDateTime;
 import java.util.Collection;
 import java.util.List;
@@ -21,7 +22,7 @@ public class Menu {
 
   private String name;
 
-  private BigDecimal price;
+  private BigInteger price;
 
   private String image;
 
@@ -33,12 +34,12 @@ public class Menu {
 
   private List<MenuOptionGroup> groups;
 
-  public BigDecimal calculateTotalPrice(int amount) {
+  public BigInteger calculateTotalPrice(int amount) {
     if (amount <= 0) {
       throw new BadInputException("주문 상품의 개수는 0이나 음수가 될 수 없습니다");
     }
 
-    return price.multiply(BigDecimal.valueOf(amount));
+    return price.multiply(BigInteger.valueOf(amount));
   }
 
   public List<MenuOption> getMenuOptions() {
