+++ src/test/java/com/flab/commerce/domain/price/PriceMapperTest.java
@@ -2,7 +2,7 @@ package com.flab.commerce.domain.price;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import org.junit.jupiter.api.Test;
 import org.mybatis.spring.boot.test.autoconfigure.MybatisTest;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,7 +19,7 @@ class PriceMapperTest {
   @Test
   void 가격을_저장한다() {
     Price pay = Price.builder()
-        .price(BigDecimal.valueOf(1_000L))
+        .price(BigInteger.valueOf(1_000L))
         .orderId(1L)
         .build();
